Update build.xml
Fix for Build Issues

Update build.xml

Modern versions of Java changed the way this is handled, and also
changed the "minimal" version acceptable, this is a temporary fix
to allow micropolis-java to build agains modern java versions.
Is still a "Deprecated" way to fix it but at leas now it builds.